push 4

+++ src/entidades/Entidad.java
@@ -26,7 +26,7 @@ public class Entidad {
     }
 
     //crea el objeto de rectangulo que servira para detectar impactos
-    protected void iniciarHitbox(float x, float y, float width, float height) {
+    protected void iniciarHitbox(float x, float y, int width, int height) {
         hitbox = new Rectangle2D.Float(x, y, width, height);
     }
